feat(grammar): implement the body replacer
Still WIP.
The body replacer, called solve_terminals_dependencies in the code,
transform a dependency from:
S -> A | '_'
A -> b

to
S -> b | '_'
A -> b

effectively removing recursion in the lexer rules.

The core idea of the function is there, but I still need to debug it
and fix the problems (too tired to focus rn). But I'm pushing this code
to have a starting point in case I will make disasters tomorrow