refactor(grammar): add len method for term/nonterm
It was useful to have a method len for terminals and non terminals in
order to avoid using the iterator and consuming it with count just to
have the number of terminals or non-terminals.

For this reason this commit refactor the class by adding the len_term()
and len_nonterm() methods.

the method is_empty() is not implemented neither for terminals nor for
non-terminals, as it can be easily derived from len_term() and
len_nonterm().

Moreover, I explicitly don't want to expose terminal or non-terminal as
I may change the underlying implementation, so only an iterator will
ever be exposed.