feat(lexer): generate regex parse tree
Added a function to generate the parse tree of a particular regex. This
is needed to generate the NFA with Thompson's algorithm. The function
itself if based on Djkstra's shunting-yard algorithm, with added support
for unary operators (most of the regex ones) both as prefix or suffix of
an operand.

The entire public interface is not stable, so tests are temporary and
nothing is documented yet.