feat(lexer): add method to canonicalise regex
Regex legal within ANTLR grammars are "augmented" and supports operators
like +, ? and ~ that are not canonical. The new method resolves them. In
order to support unicode, a new special character has been added that
represents "every character not in the alphabet". This character has
been used properly while replacing ~ and ., respectively the set
negation and the "any character" symbol.