perf(lexer): unique set literals (NFA)
Previously a lot of edges were generated while creating NFA/DFAs due to
the fact that sets like [a-z] did not account for the fact that some
transitions can be grouped: some symbols appears always together in a
set and thus can be treated as a single symbol.

For this reason the symbol table assigns unique indices to these kind of
symbols before building the automata. This has been implemented for NFAs
in this commit (DFA generation has been commented out, as it's a WIP)
and so far these are the results on the c_grammar.txt file:

NFA (previously):
- 31921 nodes, 39550 edges, 0.177s build time (release)
NFA (with this commit):
- 24041 nodes, 29700 edges, 0.125s build time (release)