perf(lexer): unique set literals (DFA)
Previously a lot of edges were generated while creating NFA/DFAs due to
the fact that sets like [a-z] did not account for the fact that some
transitions can be grouped: some symbols appears always together in a
set and thus can be treated as a single symbol.

For this reason the symbol table assigns unique indices to these kind of
symbols before building the automata. This has been implemented for DFAs
in this commit with these are the results on the c_grammar.txt file:

min DFA (previously):
- 408 nodes, 18351 edges, 0.764s build time (release)
min DFA (with this commit):
- 408 nodes, 13654 edges, 0.588s build time (release)