refactor: dynamic epsilon id + maintenance
Epsilon ID (formerly called value) is not fixed anymore but changes
depending on the symbol table.
Now the entire NFA and DFA construction is based solely on the given
symbol table and does not use fixed constants.

In addition, a general maintenance has been performed in this commit:
doc rewriting and method renaming.