revert: u8 lexer dropped in favour of 32 bit
This reverts commit c904c160ac52ab91e40b32012d9c5e7e6f4f7b9e.
In that commit an 8bit lexer reading one byte at time was explored.
In particular, while it is possible to represent certain character
classes only with single bytes sequences, others (like the any character .)
would lead to a wrong parsing unless ALL the unicode characters are
embedded into the lexer.

The reverted commit introduced the conversion of non-ASCII characters
into sequence of bytes.