refactor(simulator): buffer handling
When the lexeme is too big, buffers are extended istead of being
swapped. In addition the buffers are now on the heap instead of the
stack, and are smaller in size (4K vs 16K).

Completely rewritten the next_char routine, now it's way more robust and
easier to read, with less repeated code and branching.