feat: BSTree using vectors instead of Box
The tree is still binary (mostly) but instead of having left and right,
with option and boxes, it now has an iterable vector.

AND nodes still have a maximum of two values, OR nodes can have more
than two. This discrepancy is due to the fact that handling followpos is
slightly harder in AND nodes, and I want to keep the code as close as
possible to the dragon book.